Fixed invalid ID seg fault

+++ RoutingTable.cc
@@ -5,6 +5,7 @@
 #include <stdint.h>
 #include <iostream>
 #include <stdio.h>
+#include <cmath>
 #include <fstream>
 
 using namespace std;
@@ -36,7 +37,9 @@ Triple* RoutingTable::createTriple(uint32_t id, uint32_t address) {
 //Post: RV = true if id is in the table, otherwise false
 bool RoutingTable::isNodeInTable(uint32_t id) {
   int nthBucket = findKBucket(id);
-  bool inTable = table[nthBucket].isNodeInBucket(id);
+	bool inTable = false;
+	if(nthBucket != -1)
+  	 inTable = table[nthBucket].isNodeInBucket(id);
   return (inTable);
 }
 
@@ -70,6 +73,8 @@ bool RoutingTable::isFull() {
 //Pre: id is some valid node or key
 //Post: RV = the nth kBucket such that d = findDist(id) where 2^n <= d < 2^n+1
 int RoutingTable::findKBucket(uint32_t id) {
+	if(id > pow(2, NUMBITS) - 1)
+		return -1;
   uint32_t myDist = findDist(myId, id);
   uint32_t tempt = myDist;
   int twoPower = 0;
@@ -102,16 +107,28 @@ KBucket& RoutingTable::operator [] (int index) {
 //           false otherwise
 bool RoutingTable::addNode(uint32_t node, uint32_t address) {
   int nthBucket = findKBucket(node);
-  KBucket* currBucket = &(table[nthBucket]);
-  bool added = false;
-  Triple* newTriple = createTriple(node, address);
+	bool added = false;
 
-  if(currBucket->getNumTriples() < K){
-    //ASSERT: the bucket is not full, add to it
-    log(nthBucket, *newTriple, true);
-    currBucket->addNode(newTriple);
-    added = true;
-  }
+	if(nthBucket != -1)
+	{
+		KBucket* currBucket = &(table[nthBucket]);
+		Triple* newTriple = createTriple(node, address);
+
+		if(currBucket->getNumTriples() < K){
+			//ASSERT: the bucket is not full, add to it
+			log(nthBucket, *newTriple, true);
+			currBucket->addNode(newTriple);
+			added = true;
+		}
+	}
+	else
+	{
+		string s = "Invalid Node ID, cannot be stored in ";
+		char temp[50];
+		sprintf(temp, "%u", NUMBITS);
+		s+= string(temp) + " bits\n";
+		logError(s);
+	}
 
   return (added);
 }
@@ -120,7 +137,11 @@ bool RoutingTable::addNode(uint32_t node, uint32_t address) {
 //Post: Removes the respected Triple from the table
 void RoutingTable::deleteNode(uint32_t nodeID) {
   int nthBucket = findKBucket(nodeID);
-  table[nthBucket].deleteNode(nodeID);
+	if(nthBucket != -1)
+	{
+		log(nthBucket, nodeID);
+		table[nthBucket].deleteNode(nodeID);
+	}
 }
 
 //Pre: nodeID is a valid id, address is where nodeID is from
@@ -129,15 +150,19 @@ void RoutingTable::deleteNode(uint32_t nodeID) {
 //      RV = true if success, false otherwise
 bool RoutingTable::updateTable(uint32_t nodeID, uint32_t address) {
   int nthBucket = findKBucket(nodeID);
-  bool inTable = isNodeInTable(nodeID);
-  bool success;
-  if (inTable) {
-    table[nthBucket].adjustNode(nodeID);
-    success = true;
-  }
-  else {
-	    success = addNode(nodeID, address);
-  }
+	bool success = false;
+
+	if (nthBucket != -1)
+	{
+		bool inTable = isNodeInTable(nodeID);
+		if (inTable) {
+			table[nthBucket].adjustNode(nodeID);
+			success = true;
+		}
+		else {
+				success = addNode(nodeID, address);
+		}
+	}
   return (success);
 }
 
@@ -161,12 +186,17 @@ int RoutingTable::getKClosestNodes(uint32_t target, Triple* closeNodes) {
 //Post: RV = a copy of the oldest node in k
 Triple RoutingTable::getOldestNode(uint32_t id) {
   int nthBucket = findKBucket(id);
-  Triple* leastRecentNode = table[nthBucket].getHead();
-  Triple copy;
-  copy.node = leastRecentNode->node;
-  copy.address = leastRecentNode->address;
-  copy.port = leastRecentNode->port;
-  return (copy);
+	if(nthBucket != -1)
+	 {
+		Triple* leastRecentNode = table[nthBucket].getHead();
+		Triple copy;
+		copy.node = leastRecentNode->node;
+		copy.address = leastRecentNode->address;
+		copy.port = leastRecentNode->port;
+
+		return (copy);
+	 }
+	else return Triple();
 }
 
 //----------------------Print Table------------------------------
@@ -193,4 +223,23 @@ void RoutingTable::log(int &i, Triple curNode, bool add)
 	}
 	
 	logOut <<"port :"<< curNode.port <<" \t node : " << curNode.node << " \t address: "<< curNode.address <<endl;
+}
+
+void RoutingTable::log(int &i, uint32_t nodeID, bool add)
+{
+	if(add)
+	{
+		logOut<<"Inserted Node - ";
+	}
+	else
+	{
+		logOut<<"Deleted Node - ";
+	}
+	
+	logOut <<"port : ---- \t node : " << nodeID << " \t address: ----"<<endl;
+}
+
+void RoutingTable::logError(string msg)
+{
+	logOut << "ERROR :  "<< msg;
 }
